Move Fields to Extensible
